refactor: commonize source-media grid into SourceMediaGrid component
- Create packages/ui/src/source-media-grid.tsx with optional virtualization,
  context menu, result count, empty state, and loading state
- Replace SourceMediaScreen renderGrid prop with renderItem + enableVirtualization
- Delete server-side MediaGrid component; slim down source-media-page to thin wrapper
- Remove ~245 lines of inline grid/virtualization from tauri source-media-page
- Update parity wiki: Routes ~88% → ~95%, Components ~85% → ~92%

All type checks and tests pass.
