+++ apps/tauri/src/routes/sources/$mediaSourceId/components/source-media-page.tsx
@@ -1,31 +1,12 @@
-import {
-	ContextMenu,
-	ContextMenuContent,
-	ContextMenuItem,
-	ContextMenuSeparator,
-	ContextMenuTrigger,
-} from "@solid-imager/ui/context-menu";
 import type { MediaSourceEventTransport } from "@solid-imager/ui/hooks/use-media-source-events";
 import { useSourceMediaPage } from "@solid-imager/ui/hooks/use-source-media-page";
 import { Progress } from "@solid-imager/ui/progress";
-import {
-	SourceMediaScreen,
-	type SourceMediaScreenProps,
-} from "@solid-imager/ui/screens/source-media-screen";
+import { SourceMediaScreen } from "@solid-imager/ui/screens/source-media-screen";
 import { SearchControlPanel } from "@solid-imager/ui/search-control-panel";
 import { createQuery, useQueryClient } from "@tanstack/solid-query";
 import { useParams } from "@tanstack/solid-router";
-import { createWindowVirtualizer } from "@tanstack/solid-virtual";
 import { listen } from "@tauri-apps/api/event";
-import {
-	createEffect,
-	createMemo,
-	createSignal,
-	For,
-	onCleanup,
-	onMount,
-	Show,
-} from "solid-js";
+import { createMemo, Show } from "solid-js";
 import { MediaGridItem } from "~/components/media/media-grid-item";
 import { MoveCopyMediaDialog } from "~/components/media/move-copy-media-dialog";
 import { UploadMediaModal } from "~/components/upload-media-modal";
@@ -179,251 +160,14 @@ export function SourceMediaPage() {
 		onThumbnailReady: notifyThumbnailReady,
 	});
 
-	// --- Virtual grid setup ---
-	const MEDIA_ITEMS_PER_PAGE = 100;
-	const GRID_GAP_PX = 16;
-	const GRID_ITEM_ASPECT_RATIO = 4 / 3;
-	const VIRTUAL_ROWS_OVERSCAN = 4;
-
-	const [windowWidth, setWindowWidth] = createSignal(0);
-	const [mediaGridWidth, setMediaGridWidth] = createSignal(0);
-	let mediaGridRef: HTMLDivElement | undefined;
-
-	const columnCount = createMemo(() => {
-		const width = windowWidth();
-		if (width >= 1024) return 5;
-		if (width >= 768) return 3;
-		return 2;
-	});
-
-	const mediaItemWidth = createMemo(() => {
-		const width = mediaGridWidth();
-		const columns = columnCount();
-		if (!(width > 0 && columns > 0)) return 0;
-		return Math.max((width - GRID_GAP_PX * (columns - 1)) / columns, 0);
-	});
-
-	const mediaItemHeight = createMemo(() => {
-		const width = mediaItemWidth();
-		if (width <= 0) return 0;
-		return width * GRID_ITEM_ASPECT_RATIO;
-	});
-
-	const mediaRows = createMemo(() => {
-		const results = page.mediaResults();
-		const columns = columnCount();
-		const rows: (typeof results)[] = [];
-		for (let index = 0; index < results.length; index += columns) {
-			rows.push(results.slice(index, index + columns));
-		}
-		return rows;
-	});
-
-	const rowCount = createMemo(() => mediaRows().length);
-
-	const mediaRowVirtualizer = createWindowVirtualizer<HTMLDivElement>({
-		get count() {
-			return rowCount();
-		},
-		estimateSize: () => mediaItemHeight() || 320,
-		gap: GRID_GAP_PX,
-		getItemKey: (index) => index,
-		overscan: VIRTUAL_ROWS_OVERSCAN,
-		scrollMargin: 0,
-	});
-
-	const useVirtualGrid = createMemo(
-		() =>
-			page.mediaResults().length > MEDIA_ITEMS_PER_PAGE && mediaItemWidth() > 0,
-	);
-
-	const updateMediaGridMetrics = () => {
-		if (!mediaGridRef) return;
-		setMediaGridWidth(mediaGridRef.getBoundingClientRect().width);
-	};
-
-	onMount(() => {
-		setWindowWidth(window.innerWidth);
-		updateMediaGridMetrics();
-
-		const handleResize = () => {
-			setWindowWidth(window.innerWidth);
-			updateMediaGridMetrics();
-		};
-		window.addEventListener("resize", handleResize);
-
-		const resizeObserver = new ResizeObserver(() => {
-			updateMediaGridMetrics();
-		});
-		if (mediaGridRef) {
-			resizeObserver.observe(mediaGridRef);
-		}
-
-		onCleanup(() => {
-			window.removeEventListener("resize", handleResize);
-			resizeObserver.disconnect();
-		});
-	});
-
-	createEffect(() => {
-		rowCount();
-		mediaItemHeight();
-		columnCount();
-		mediaRowVirtualizer.measure();
-	});
-
-	createEffect(() => {
-		const virtualRows = mediaRowVirtualizer.getVirtualItems();
-		const lastRow = virtualRows[virtualRows.length - 1];
-		if (!lastRow) return;
-		if (
-			lastRow.index >= rowCount() - 2 &&
-			page.mediaQuery.hasNextPage &&
-			!page.mediaQuery.isFetchingNextPage
-		) {
-			void page.mediaQuery.fetchNextPage();
-		}
-	});
-
-	// --- Render props ---
-	const renderGrid: SourceMediaScreenProps["renderGrid"] = (gridProps) => (
-		<div class="min-h-0 space-y-4">
-			<div class="mb-4 flex items-center justify-between">
-				<p class="text-gray-600 text-sm">
-					{page.mediaQuery.data?.pages[0]?.total ?? 0} 件の結果
-				</p>
-			</div>
-
-			<ContextMenu>
-				<ContextMenuTrigger class="block w-full">
-					<div
-						class="relative w-full"
-						ref={(element) => {
-							mediaGridRef = element;
-							requestAnimationFrame(() => {
-								updateMediaGridMetrics();
-							});
-						}}
-						style={{
-							height: useVirtualGrid()
-								? `${mediaRowVirtualizer.getTotalSize()}px`
-								: undefined,
-						}}
-					>
-						<Show
-							fallback={
-								<div class="grid grid-cols-2 gap-4 md:grid-cols-3 lg:grid-cols-5">
-									<For each={gridProps.mediaResults()}>
-										{(media) => (
-											<MediaGridItem
-												media={media}
-												onContextMenu={() =>
-													gridProps.setContextMenuMediaId(media.id)
-												}
-												sourceRootPath={sourceRootPath()}
-											/>
-										)}
-									</For>
-								</div>
-							}
-							when={useVirtualGrid()}
-						>
-							<For each={mediaRowVirtualizer.getVirtualItems()}>
-								{(virtualRow) => {
-									const rowMedia = () => mediaRows()[virtualRow.index] || [];
-									return (
-										<div
-											class="absolute left-0 top-0 grid gap-4"
-											style={{
-												"grid-template-columns": `repeat(${columnCount()}, minmax(0, 1fr))`,
-												height: `${virtualRow.size}px`,
-												transform: `translateY(${virtualRow.start}px)`,
-												width: "100%",
-											}}
-										>
-											<For each={rowMedia()}>
-												{(media) => (
-													<MediaGridItem
-														media={media}
-														onContextMenu={() =>
-															gridProps.setContextMenuMediaId(media.id)
-														}
-														sourceRootPath={sourceRootPath()}
-													/>
-												)}
-											</For>
-										</div>
-									);
-								}}
-							</For>
-						</Show>
-					</div>
-				</ContextMenuTrigger>
-				<ContextMenuContent>
-					<Show
-						fallback={
-							<ContextMenuItem disabled>No media selected</ContextMenuItem>
-						}
-						when={gridProps.contextMenuMediaId()}
-					>
-						<ContextMenuItem
-							onSelect={() => {
-								const id = gridProps.contextMenuMediaId();
-								if (id) gridProps.onDelete(id);
-							}}
-						>
-							Delete
-						</ContextMenuItem>
-						<ContextMenuSeparator />
-						<ContextMenuItem
-							onSelect={() => {
-								const id = gridProps.contextMenuMediaId();
-								if (id) gridProps.onCopyMove(id, "copy");
-							}}
-						>
-							Copy to Source
-						</ContextMenuItem>
-						<ContextMenuItem
-							onSelect={() => {
-								const id = gridProps.contextMenuMediaId();
-								if (id) gridProps.onCopyMove(id, "move");
-							}}
-						>
-							Move to Source
-						</ContextMenuItem>
-						<ContextMenuSeparator />
-						<ContextMenuItem
-							onSelect={() => {
-								const id = gridProps.contextMenuMediaId();
-								if (id) void gridProps.onSyncSingleMedia(id);
-							}}
-						>
-							Sync Metadata
-						</ContextMenuItem>
-					</Show>
-				</ContextMenuContent>
-			</ContextMenu>
-
-			<Show
-				when={gridProps.mediaResults().length === 0 && !gridProps.isPending}
-			>
-				<div class="py-12 text-center text-gray-500">
-					検索結果が見つかりませんでした
-				</div>
-			</Show>
-
-			<div
-				class="flex h-10 w-full items-center justify-center text-gray-500"
-				ref={gridProps.setLoadMoreRef}
-			>
-				<Show when={gridProps.isFetchingNextPage}>Loading more...</Show>
-			</div>
-		</div>
-	);
-
-	const renderActions: SourceMediaScreenProps["renderActions"] = (
-		actionProps,
-	) => (
+	const renderActions = (actionProps: {
+		isSyncing: boolean;
+		isSyncDisabled: boolean;
+		onDumpDownload: (mode?: "json" | "zip") => void;
+		onSyncLoadedMedia: () => void;
+		onAddMedia: () => void;
+		onRestore: () => void;
+	}) => (
 		<MediaListActions
 			filterPanel={
 				<SearchControlPanel
@@ -447,9 +191,11 @@ export function SourceMediaPage() {
 		/>
 	);
 
-	const renderJobProgress: SourceMediaScreenProps["renderJobProgress"] = (
-		progressProps,
-	) => (
+	const renderJobProgress = (progressProps: {
+		jobProgress: () =>
+			| import("@solid-imager/core/domain/sources/events").JobProgressEvent
+			| null;
+	}) => (
 		<Show when={progressProps.jobProgress()}>
 			{(progress) => (
 				<div class="mb-4 rounded-md border bg-muted/50 px-4 py-3">
@@ -461,7 +207,9 @@ export function SourceMediaPage() {
 							{Math.round((progress().processed / progress().total) * 100)}%
 						</span>
 					</div>
-					<Progress value={(progress().processed / progress().total) * 100} />
+					<Progress
+						value={(progress().processed / progress().total) * 100}
+					/>
 				</div>
 			)}
 		</Show>
@@ -469,9 +217,16 @@ export function SourceMediaPage() {
 
 	return (
 		<SourceMediaScreen
+			enableVirtualization
 			page={page}
 			renderActions={renderActions}
-			renderGrid={renderGrid}
+			renderItem={(media, { onContextMenu }) => (
+				<MediaGridItem
+					media={media}
+					onContextMenu={onContextMenu}
+					sourceRootPath={sourceRootPath()}
+				/>
+			)}
 			renderJobProgress={renderJobProgress}
 			renderMoveCopyDialog={() => (
 				<MoveCopyMediaDialog
